Completed all possible TODO's
