+++ backend/checks/windows/startup_test.go
@@ -3,6 +3,7 @@ package windows_test
 import (
 	"errors"
 	"github.com/InfoSec-Agent/InfoSec-Agent/backend/checks/windows"
+	"golang.org/x/sys/windows/registry"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -63,4 +64,29 @@ func TestStartup(t *testing.T) {
 			require.Equal(t, tt.want, got)
 		})
 	}
+}
+
+// TestFindEntriesInvalidInput is a test function that validates the behavior of the FindEntries function when provided with invalid (empty) input.
+//
+// Parameter:
+//   - t *testing.T: The testing framework instance used to run the test and report the results.
+//
+// This function does not return any values. It uses the testing framework to assert that the FindEntries function behaves as expected when provided with an empty list of entries and an invalid registry key. Specifically, it checks that the function returns an empty list of entries. If the FindEntries function does not behave as expected, this test function will cause the test run to fail.
+func TestFindEntriesInvalidInput(t *testing.T) {
+	key := registry.Key(0x0)
+	var entries []string
+	elements := windows.FindEntries(entries, mocking.NewRegistryKeyWrapper(key))
+	require.Empty(t, elements)
+}
+
+func TestCheckAllZero(t *testing.T) {
+	entries := []byte{0, 0, 0, 0}
+	result := windows.CheckAllZero(entries)
+	require.True(t, result)
+	entries = []byte{0, 0, 0, 1}
+	result = windows.CheckAllZero(entries)
+	require.False(t, result)
+	entries = []byte{}
+	result = windows.CheckAllZero(entries)
+	require.True(t, result)
 }
